requirements.txt
python-whois
dnspython
requests
shodan
ssl
socket

